Removed int type casting from purchase price and reducepriceBy (#3)

+++ app/code/community/Openpay/Payment/Model/paymentadapter/businessLayer/Magento/Shopsystem.php
@@ -34,7 +34,7 @@ class Shopsystem
         }
         $shopCart->total = $cart->getGrandTotal();
         $integerTotal = round((float)$cart->getGrandTotal(), 2);
-        $shopCart->integerTotal = (int)($integerTotal * 100);
+        $shopCart->integerTotal = ($integerTotal * 100);
         $shopCart->deliveryAddress = $cart->getShippingAddress()->getData();
         $shopCart->invoiceAddress = $cart->getBillingAddress()->getData();
